build(codegen): declaration of named html colors from text file
Create the build script that parses "html-color-names.txt" file during
compilation and generates constant variables for each color, along with
the array containing all colors. This makes it possible to implement
functionality that will deal with the named html colors in the future.

This is not implemented as a procedural macro, as it requires heavy
dependencies and converting the project into a Cargo workspace, thereby
adding complexity to the small project. Alternatively, we could directly
hardcode the values in the code and avoid complicating our lives with
the build script. However, I wanted to practice procedural macros and
code generation using a build script. Additionally, this approach allows
us to easily modify almost anything in the generated code without a
significant git diff.