refactor: introduce ColorDistance trait, split nix run-clippy package
Introduce "ColorDistance" trait, which will make it easier to add more
color distance algorithms. "Premature optimization is the root of all
evil" but I honestly don't care in this particular situation. I want to
have very generic code.
Create two different packages in the flake for clippy to "warn" and
"deny". Because if you use "deny", then not all lints will be shown. So
in this case, only lints from clippy::cargo appeared, but others were
ignored due to the nature of "deny".