feat: parse rgb from hex string
Add functionality to parse Rgb from strings like "#FAFA00". This
unlocks writing a CLI program. Write tests and run cargo-mutant and
cargo-tarpaulin. Though coverage is not 100% but the most tricky
functions like "Rgb::from_hex" are tested properly.
Produce two more const arrays from html "COLORS" so we can use
"ColorDistance". Previously, because we had "COLORS" as the array of
tuples, we couldn't pass it to the "find_closest" function. By splitting
it into two arrays, we can pass "ONLY_COLORS" to the function and then
use "ONLY_NAMES" to return the name.