feat: --all-html option that prints named html colors
Make uppercase rgb, xyz and lab color spaces.

Fix the issue where every algorithm shows the same result when running
binary. This was caused by variable shadowing: input was stored in the
"color" variable and we also  saved result of previous algorithm in
"color" too. Now I see that each algorithm gives a different result
(interesting), good testing will be even harder.

Fix simple bug in the RGB to XYZ conversion that resulted in incorrect
CIE94 results. It turns out that the XYZ values needed to be normalized
by multiplying by 100.

Comment colors in the html-color-names.txt file, which has the same
value but a different name. It's temporary until I find out a way how to
handle them properly.

Use "core" instead of "std" to add "no_std" flag to the library in the future.