tests: add integration tests to verify cli stdout/err
Write mini-test-framework based on the "ron" file format (since it
supports raw strings) and serde. Each test case is defined by args,
stdout, stderr, etc. Test cases are generated using the build script.
I could read from the "ron" file and execute as is, but then I would
only saw the first failed test, so I decided to struggle and create this
complex structure (and wrote 2 tests, xd). It needs more functionality
in the future, like "stdout.contains" or something similar. Because I
don't want to copy the full output of "--help" into a test case to test
it.

Rename "libcodi" to "codi-core".
Move the binary package to the separate workspace package "codi-bin".
Replace example image in the README that I forgot to do in the previous
commit.